pyproject.toml
[tool.poetry]
name = "zeus-collector"
version = "0.1.0"
description = "Data collector module for MetaProject Zeus"
authors = ["Petr Balnar <petr.balnar@seznam.cz>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
zeus-collector = "src.main:main"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.9.0"
httpx = "^0.27.0"
rich = "^13.9.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.6.5"
debugpy = "^1.8.17"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"